copy migrations to server before running migrations
In the case of downgrading a release this is ok. If any migrations need to be downgraded, they did not exist in the previous commits
therefore they will not be overwritten by the copy. This allows both downgrades and upgrades to run correctly. In the future, in the case of
a downgrade, maybe they will be removed by another step after the migration to clean up the server file system and prevent confusion.